Fix the spaghetti code from the Librarian Dashboard (it got separated into 5 files: LibrarianOrders, LibrarianReaders, LibrarianReturns, LibrarianSettings and LibrarianHomePage).
