+++ web_app/src/App.tsx
@@ -3,21 +3,33 @@ import './index.css';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 import LandingPage from './LandingPage';
-import LibrarianLogin from './Librarian/LibrarianLogin.tsx';
-import LibraryAdminLogin from './LibraryAdmin/LibraryAdminLogin.tsx';
+
+// System Administrator
 import SysAdminLogin from "./SystemAdmin/SystemAdminLogin.tsx";
-import RegistrationForm from './LibraryAdmin/LibraryAdminRegisterForm.tsx';
-import ProcessingPage from './Utils/ProcessingInfoPage.tsx';
 import SystemAdminDashboard from './SystemAdmin/SystemAdminHomePage';
-import LibraryAdminAddLibrary from './LibraryAdmin/LibraryAdminAddLibrary';
-import LibraryAdminHomePage from './LibraryAdmin/LibraryAdminHomePage.tsx';
 import SubmissionDetailsLibrary from './SystemAdmin/SubmissionDetailsLibrary.tsx';
 import SubmissionDetailsDriver from './SystemAdmin/SubmissionDetailsDriver.tsx';
-import LibrarianHomePage from './Librarian/LibrarianHomePage';
+
+// Library Administrator
+import RegistrationForm from './LibraryAdmin/LibraryAdminRegisterForm.tsx';
+import LibraryAdminLogin from './LibraryAdmin/LibraryAdminLogin.tsx';
+import LibraryAdminHomePage from './LibraryAdmin/LibraryAdminHomePage.tsx';
+import LibraryAdminAddLibrary from './LibraryAdmin/LibraryAdminAddLibrary';
+
+// Librarian
+import LibrarianLogin from './Librarian/LibrarianLogin.tsx';
+import LibrarianHomePage from './Librarian/LibrarianHomePage.tsx';
+import LibrarianAddBook from './Librarian/LibrarianAddBook.tsx';
+import LibrarianOrders from './Librarian/LibrarianOrders.tsx';
+import LibrarianReturns from './Librarian/LibrarianReturns.tsx';
+import LibrarianReaders from './Librarian/LibrarianReaders.tsx';
+import LibrarianSettings from './Librarian/LibrarianSettings.tsx';
+
+// Utils
+import ProcessingPage from './Utils/ProcessingInfoPage.tsx';
 import UserManualPage from './Utils/UserManualPage.tsx';
 import ContactInfoPage from './Utils/ContactInfoPage.tsx';
 import LegalInfoPage from './Utils/LegalInfoPage.tsx';
-import AddBook from './Librarian/AddBook.tsx';
 
 import SwaggerUI from 'swagger-ui-react';
 import 'swagger-ui-react/swagger-ui.css';
@@ -41,7 +53,11 @@ const App: React.FC = () => {
                 <Route path="/contact" element={<ContactInfoPage />} />
                 <Route path="/legal-info" element={<LegalInfoPage />} />
                 <Route path="/library-admin-dashboard" element={<LibraryAdminHomePage />} />
-                <Route path="/add-book" element={<AddBook />} />
+                <Route path="/add-book" element={<LibrarianAddBook />} />
+                <Route path="/orders" element={<LibrarianOrders />} />
+                <Route path="/returns" element={<LibrarianReturns />} />
+                <Route path="/readers" element={<LibrarianReaders />} />
+                <Route path="/librarian-settings" element={<LibrarianSettings />} />
                 <Route
                     path="/api-docs"
                     element={
